feat: no longer use logical CPU as concurrency

+++ internal/download/command.go
@@ -11,12 +11,11 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
-	"runtime"
 	"strings"
 )
 
 type Command struct {
-	MaxConcurrency int `short:"P" long:"max-concurrency" description:"use up to max-concurrency number of goroutines at a time. If not given, default to the number of logical CPUs." default:"0"`
+	MaxConcurrency int `short:"P" long:"max-concurrency" description:"use up to max-concurrency number of goroutines at a time for range downloads." default:"5"`
 	Args           struct {
 		Files []flags.Filename `positional-arg-name:"file" description:"the local files each containing a single S3 URI" required:"yes"`
 	} `positional-args:"yes"`
@@ -29,14 +28,8 @@ func (c *Command) Execute(args []string) error {
 		return fmt.Errorf("unknown positional arguments: %s", strings.Join(args, " "))
 	}
 
-	switch {
-	case c.MaxConcurrency < 0:
-		return fmt.Errorf("max-concurrency cannot be negative")
-	case c.MaxConcurrency == 0:
-		c.MaxConcurrency = runtime.NumCPU()
-		log.Printf("using max concurrency %d (logical CPU count)", c.MaxConcurrency)
-	default:
-		log.Printf("using max concurrency %d", c.MaxConcurrency)
+	if c.MaxConcurrency <= 0 {
+		return fmt.Errorf("max-concurrency must be positive")
 	}
 
 	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, os.Kill)
